* Add `Makefile` to replace `build.sh` script.
-This is better, and allows support of more build-related functionalities down the road (including system tests and linting scripts).

* Add all make targets to `PHONY`.

* Update `master` README's `Building` section.

* Add compiler options to enable warnings/errors.
-Add `-Werror` to turn all compiler warnings into errors.

* Fix compilation errors.
-Arose because of recent addition of `-Werror` and other compile options.

+++ Software/main/main.cpp
@@ -24,7 +24,7 @@ int main() {
     dataPackets[0] = new spartan::IMUDataPacket;
 
     // Initialization of sensors
-    for (int i = 0; i < sensors.size(); i++) {
+    for (size_t i = 0; i < sensors.size(); i++) {
 	    sensors[i]->powerOn();
     }
     std::ofstream fout;
@@ -32,14 +32,14 @@ int main() {
 
     // TODO: flight loop
     for (int count  = 0; count < 100; count++) {
-        for (int i = 0; i < sensors.size(); i++) {
+        for (size_t i = 0; i < sensors.size(); i++) {
             mdp.timestamp = spartan::getTimeMillis();
             if (sensors[i]->pollData(mdp) != spartan::RESULT_SUCCESS) {
                 std::cerr << "Sensor name: " << sensors[i]->name() << " instance; " << sensors[i]->getInstance()
                     << " poll data error!" << std::endl;
             }
         }
-        for (int i = 0; i < dataPackets.size(); i++) {
+        for (size_t i = 0; i < dataPackets.size(); i++) {
             dataPackets[i]->populate(mdp);
             if (DEBUG) {
                 std::cout << "Packet size " << dataPackets[i]->getSize() << std::endl;
